// client/src/model/enums/lefthandlayout.ts
export enum LeftHandLayout {
  _2_BS_ROWS = "_2_BS_ROWS",
  _3_BS_ROWS_A_7TH = "_3_BS_ROWS_A_7TH",
  _3_BS_ROWS_A_5_DIM = "_3_BS_ROWS_A_5_DIM",
  _3_BS_ROWS_B_7TH = "_3_BS_ROWS_B_7TH",
  _3_BS_ROWS_B_5_DIM = "_3_BS_ROWS_B_5_DIM",
  _3_BS_ROWS_BX_7TH_A = "_3_BS_ROWS_BX_7TH_A",
  _3_BS_ROWS_BX_7TH_B = "_3_BS_ROWS_BX_7TH_B",
  _3_BS_ROWS_BELGIUM = "_3_BS_ROWS_BELGIUM",
  PIANO_LEFT_HAND = "PIANO_LEFT_HAND",
}